Recombined the report types
